pyproject.toml
[project]
name = "shows"
version = "0.0.0"
description = ""
authors = [{name = "Anton Samarchyan", email = "desecho@gmail.com"}]
license = {text = "MIT"}
readme = "README.rst"
requires-python = ">=3.13"

dependencies = [
    # Django
    "Django>=5.2.4,<6",
    "django-annoying>=0.10.6,<1",
    "django-tqdm>=1.3.1,<2",
    "django-custom-anonymous>=0.5.0,<1",
    "django-modeladmin-reorder>=0.3.1,<1",
    "django-stubs-ext>=5.2.2,<6",
    "django-celery-results>=2.4.0,<3",
    "django-countries[pyuca]>=7.6.1,<8",
    "django-timezone-field>=7.1,<8",
    "djangorestframework>=3.13.1,<4",
    "djangorestframework-simplejwt>=5.5.1,<6",
    "django-cors-headers>=4.4.0,<5",
    "django-rest-registration>=0.9.0,<1",

    # Requests
    "requests>=2.31.0,<3",
    "requests-toolbelt>=1.0.0,<2",

    # Dates
    "python-dateutil>=2.8.2,<3",
    "Babel>=2.10.3,<3",

    "mysqlclient>=2.2.0,<3",
    "sentry-sdk>=2.14.0,<3",
    "gunicorn>=23.0.0,<24",
    "typing-extensions>=4.3.0,<5",
    "Authlib>=1.0.1,<2",
    "pip>=25.2",
    "wget>=3.2,<4",
    "tmdbsimple>=2.9.1,<3",
    "celery[redis]>=5.2.7,<6",
    "tzdata>=2025.2",

    # AI/ML
    "openai>=1.0.0,<2",
]

[dependency-groups]
dev = [
    # Pytest
    "pytest",
    "pytest-cov",
    "pytest-sugar",
    "pytest-mock",
    "pytest-django",

    # Django
    "django-test-without-migrations",

    # Linting
    "mypy",

    # Stubs
    "django-stubs>=5.2.2,<6",
    "djangorestframework-stubs",
    "types-python-dateutil",
    "types-requests",
    "types-babel",
    "celery-types",

    "beautifulsoup4",
    "requests-mock",
    "ruff",
]

[tool.hatch.build.targets.wheel]
packages = ["src/shows"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
